requirements.txt
# Core
gradio==4.31.5
google-generativeai==0.5.4

# Vector DB (file-based, no server needed)
chromadb==0.5.0

# Relational DB
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Document parsing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0
youtube-transcript-api==0.6.2

# Auth / security
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
Pillow==10.3.0
